requirements.txt
beautifulsoup4==4.14.2
geopy==2.4.1
groq==0.32.0
langchain_core==0.3.77
langchain_groq==0.3.8
networkx==3.1
numpy==1.24.4
openai==2.1.0
opencage==3.2.0
openrouteservice==2.3.3
piexif==1.1.3
Pillow==11.3.0
progress==1.6.1
python-dotenv==1.1.1
qdrant_client==1.15.1
Requests==2.32.5
scikit_learn==1.7.2
sentence_transformers==5.1.1
serpapi==0.1.5
st_copy_to_clipboard==0.1.6
streamlit==1.50.0
streamlit_modal==0.1.2
